update createMathOperation.ts
Here we can use one or instead of three ands to reduce the cognitive complexity due to the number of if-else structures. Since all these three cases of and can be written as one or. After that, we use the Nullish Coalescing Operator (??) for the two arguments and the default value as an escape hatch.